Move `mk-tmp-dir-in` to `tests/helpers`
