+++ tests/acceptance/addon-dummy-generate-test.js
@@ -7,7 +7,7 @@ let root = process.cwd();
 let tmproot = path.join(root, 'tmp');
 const Blueprint = require('../../lib/models/blueprint');
 const BlueprintNpmTask = require('ember-cli-internal-test-helpers/lib/helpers/disable-npm-on-blueprint');
-const mkTmpDirIn = require('../../lib/utilities/mk-tmp-dir-in');
+const mkTmpDirIn = require('../helpers/mk-tmp-dir-in');
 
 const { expect } = require('chai');
 const { file } = require('chai-files');
