+++ tests/integration/tasks/build-test.js
@@ -9,7 +9,7 @@ const BuildTask = require('../../../lib/tasks/build');
 const MockProject = require('../../helpers/mock-project');
 const MockProcess = require('../../helpers/mock-process');
 const copyFixtureFiles = require('../../helpers/copy-fixture-files');
-const mkTmpDirIn = require('../../../lib/utilities/mk-tmp-dir-in');
+const mkTmpDirIn = require('../../helpers/mk-tmp-dir-in');
 const willInterruptProcess = require('../../../lib/utilities/will-interrupt-process');
 let root = process.cwd();
 let tmproot = path.join(root, 'tmp');
