+++ tests/unit/models/builder-test.js
@@ -6,7 +6,7 @@ const BuildCommand = require('../../../lib/commands/build');
 const commandOptions = require('../../factories/command-options');
 const fixturify = require('fixturify');
 const MockProject = require('../../helpers/mock-project');
-const mkTmpDirIn = require('../../../lib/utilities/mk-tmp-dir-in');
+const mkTmpDirIn = require('../../helpers/mk-tmp-dir-in');
 const td = require('testdouble');
 const { expect } = require('chai');
 const { file } = require('chai-files');
